Adding function stubs for conversion functions
