+++ elfpy/data/db_schema.py
@@ -1,7 +1,6 @@
 """Pool info struct retured from hyperdrive contract"""
 from __future__ import annotations
 
-import logging
 from datetime import datetime
 from typing import Any, Union
 
@@ -9,11 +8,6 @@ from eth_typing import BlockNumber
 from hexbytes import HexBytes
 from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import DeclarativeBase, Mapped, MappedAsDataclass, mapped_column
-from web3 import Web3
-from web3.contract.contract import Contract
-from web3.types import BlockData
-
-from elfpy import eth, hyperdrive_interface
 
 # Schema file doesn't need any methods in these dataclasses
 # pylint: disable=too-few-public-methods
@@ -237,155 +231,3 @@ class UserMap(Base):
     # autoincrement with BigIntegers. Hence, we use the Integer variant when using sqlite in tests
     address: Mapped[str] = mapped_column(String, primary_key=True)
     username: Mapped[str] = mapped_column(String, index=True)
-
-
-def fetch_transactions_for_block(web3: Web3, contract: Contract, block_number: BlockNumber) -> list[Transaction]:
-    """Fetch transactions related to the contract.
-
-    Returns the block pool info from the Hyperdrive contract
-
-    Arguments
-    ---------
-    web3: Web3
-        web3 provider object
-    contract: Contract
-        The contract to query the pool info from
-    block_number: BlockNumber
-        The block number to query from the chain
-
-    Returns
-    -------
-    list[Transaction]
-        A list of Transaction objects ready to be inserted into Postgres
-    """
-    block: BlockData = web3.eth.get_block(block_number, full_transactions=True)
-    transactions = block.get("transactions")
-    if not transactions:
-        logging.info("no transactions in block %s", block.get("number"))
-        return []
-    out_transactions = []
-    for transaction in transactions:
-        if isinstance(transaction, HexBytes):
-            logging.warning("transaction HexBytes")
-            continue
-        if transaction.get("to") != contract.address:
-            logging.warning("transaction not from contract")
-            continue
-        transaction_dict: dict[str, Any] = dict(transaction)
-        # Convert the HexBytes fields to their hex representation
-        tx_hash = transaction.get("hash") or HexBytes("")
-        transaction_dict["hash"] = tx_hash.hex()
-        # Decode the transaction input
-        try:
-            method, params = contract.decode_function_input(transaction["input"])
-            transaction_dict["input"] = {"method": method.fn_name, "params": params}
-        except ValueError:  # if the input is not meant for the contract, ignore it
-            continue
-        tx_receipt = web3.eth.get_transaction_receipt(tx_hash)
-        logs = eth.get_transaction_logs(web3, contract, tx_receipt)
-        receipt: dict[str, Any] = _recursive_dict_conversion(tx_receipt)  # type: ignore
-        out_transactions.append(_build_transaction_object(transaction_dict, logs, receipt))
-    return out_transactions
-
-
-def _build_transaction_object(
-    transaction_dict: dict[str, Any],
-    logs: list[dict[str, Any]],
-    receipt: dict[str, Any],
-) -> Transaction:
-    """Conversion function to translate output of chain queries to the Transaction object.
-
-    Arguments
-    ----------
-    transaction_dict : dict[str, Any]
-        A dictionary representing the decoded transactions from the query
-    logs: list[str, Any]
-        A dictionary representing the decoded logs from the query
-    receipt: dict[str, Any]
-        A dictionary representing the transaction receipt from the query
-
-    Returns
-    -------
-    Transaction
-        A transaction object to be inserted into postgres
-    """
-    # Build output obj dict incrementally to be passed into Transaction
-    # i.e., Transaction(**out_dict)
-    # Base transaction fields
-    out_dict: dict[str, Any] = {
-        "blockNumber": transaction_dict["blockNumber"],
-        "transactionIndex": transaction_dict["transactionIndex"],
-        "nonce": transaction_dict["nonce"],
-        "transactionHash": transaction_dict["hash"],
-        "txn_to": transaction_dict["to"],
-        "txn_from": transaction_dict["from"],
-        "gasUsed": receipt["gasUsed"],
-    }
-    # Input solidity methods and parameters
-    # TODO can the input field ever be empty or not exist?
-    out_dict["input_method"] = transaction_dict["input"]["method"]
-    input_params = transaction_dict["input"]["params"]
-    out_dict["input_params_contribution"] = eth.convert_scaled_value(input_params.get("_contribution", None))
-    out_dict["input_params_apr"] = eth.convert_scaled_value(input_params.get("_apr", None))
-    out_dict["input_params_destination"] = input_params.get("_destination", None)
-    out_dict["input_params_asUnderlying"] = input_params.get("_asUnderlying", None)
-    out_dict["input_params_baseAmount"] = eth.convert_scaled_value(input_params.get("_baseAmount", None))
-    out_dict["input_params_minOutput"] = eth.convert_scaled_value(input_params.get("_minOutput", None))
-    out_dict["input_params_bondAmount"] = eth.convert_scaled_value(input_params.get("_bondAmount", None))
-    out_dict["input_params_maxDeposit"] = eth.convert_scaled_value(input_params.get("_maxDeposit", None))
-    out_dict["input_params_maturityTime"] = input_params.get("_maturityTime", None)
-    out_dict["input_params_minApr"] = eth.convert_scaled_value(input_params.get("_minApr", None))
-    out_dict["input_params_maxApr"] = eth.convert_scaled_value(input_params.get("_maxApr", None))
-    out_dict["input_params_shares"] = eth.convert_scaled_value(input_params.get("_shares", None))
-    # Assuming one TransferSingle per transfer
-    # TODO Fix this below eventually
-    # There can be two transfer singles
-    # Currently grab first transfer single (e.g., Minting hyperdrive long, so address 0 to agent)
-    # Eventually need grabbing second transfer single (e.g., DAI from agent to hyperdrive)
-    event_logs = [log for log in logs if log["event"] == "TransferSingle"]
-    if len(event_logs) == 0:
-        event_args: dict[str, Any] = {}
-        # Set args as None
-    elif len(event_logs) == 1:
-        event_args: dict[str, Any] = event_logs[0]["args"]
-    else:
-        logging.warning("Tranfer event contains multiple TransferSingle logs, selecting first")
-        event_args: dict[str, Any] = event_logs[0]["args"]
-    out_dict["event_value"] = eth.convert_scaled_value(event_args.get("value", None))
-    out_dict["event_from"] = event_args.get("from", None)
-    out_dict["event_to"] = event_args.get("to", None)
-    out_dict["event_operator"] = event_args.get("operator", None)
-    out_dict["event_id"] = event_args.get("id", None)
-    # Decode logs here
-    if out_dict["event_id"] is not None:
-        event_prefix, event_maturity_time = hyperdrive_interface.decode_asset_id(out_dict["event_id"])
-        out_dict["event_prefix"] = event_prefix
-        out_dict["event_maturity_time"] = event_maturity_time
-    transaction = Transaction(**out_dict)
-    return transaction
-
-
-def _recursive_dict_conversion(obj: Any) -> Any:
-    """Recursively converts a dictionary to convert objects to hex values.
-
-    Arguments
-    ---------
-    obj : Any
-        Could be a HexBytes, dict, or any object with the `items` attribute
-
-    Returns
-    -------
-    Any
-        A nested dictionary containing the decoded object values
-
-
-    .. todo::
-        This function needs to be better constrained & typed, or avoided all together?
-    """
-    if isinstance(obj, HexBytes):
-        return obj.hex()
-    if isinstance(obj, dict):
-        return {key: _recursive_dict_conversion(value) for key, value in obj.items()}
-    if hasattr(obj, "items"):  # any other type with "items" attr, e.g. TypedDict and OrderedDict
-        return {key: _recursive_dict_conversion(value) for key, value in obj.items()}
-    return obj
